fourier/internal/fftpack: factor out FFTPACK port to internal package
